convert command line argument for effective population size to long double instead of double.

+++ src/pggtyper-paths.cpp
@@ -95,7 +95,7 @@ int main (int argc, char* argv[])
 	nr_core_threads = stoi(argument_parser.get_argument('t'));
 	only_genotyping = argument_parser.get_flag('g');
 	only_phasing = argument_parser.get_flag('p');
-	effective_N = stod(argument_parser.get_argument('n'));
+	effective_N = stold(argument_parser.get_argument('n'));
 
 	// print info
 	cerr << "Files and parameters used:" << endl;
